Adding useUserRegistration hook
