+++ src/screens/UserRegistration/index.js
@@ -1,38 +1,26 @@
-import React, { useState, useContext } from 'react';
-import { Link, useHistory } from 'react-router-dom';
+import React, { useState } from 'react';
+import { Link } from 'react-router-dom';
 import { TextField, Button, Grid, CircularProgress, Typography } from '@material-ui/core';
 
 import useErrors from '../../hooks/useErrors';
-import { SnackBarContext } from '../../contexts/snackbar';
 
-import { userRegistrationService } from './services';
 import { useStyles } from './styles';
 import validations from './validations';
+import useUserRegistration from './hooks/useUserRegistration';
 
 const UserRegistration = () => {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
   const classes = useStyles();
-  const { addAlert } = useContext(SnackBarContext);
-  const history = useHistory();
+
+  const { isLoading, register } = useUserRegistration();
 
   const [errors, validateFields, isFormValid] = useErrors(validations);
 
-  const onSubmit = async (event) => {
+  const onSubmit = (event) => {
     event.preventDefault();
-    setIsLoading(true);
 
-    await userRegistrationService(email, password)
-      .catch(() => {
-        addAlert({ content: 'Não foi possível criar o usuário.', customSeverity: 'error' });
-      })
-      .finally(() => {
-        setIsLoading(false);
-        history.push('/usuarios');
-      });
-
-    setIsLoading(false);
+    register(email, password);
   };
 
   const handleUpdateEmail = (event) => {
